feat: 添加oss上传成功后，是否删除本地文件参数

+++ deploy-ali-oss.js
@@ -18,10 +18,12 @@ class DeployAliOss {
       ossTimeout,
       ossNamespace,
       ossPattern,
+      ossClearLocalFile,
       versionsRetainedNumber, // 保留的版本数量
       local_target = path.resolve('dist'),
     } = config;
 
+    this.config.ossClearLocalFile = typeof ossClearLocalFile === 'boolean' ? ossClearLocalFile : true;
     this.config.ossNamespace = ossNamespace || 'frontend';
     this.config.ossPattern = ossPattern || `${local_target}/**/*.!(html)`;
     this.config.versionsRetainedNumber = Math.max(versionsRetainedNumber, 1);
@@ -114,9 +116,11 @@ class DeployAliOss {
       await this.uploadFile(item.fileName, item.filePath);
     });
     console.log('上传到OSS完成');
-    console.log('清理dist目录...');
-    await Promise.all(files.map(item => fs.remove(item)));
-    console.log('清理dist目录完成');
+    if (this.config.ossClearLocalFile) {
+      console.log('清理dist目录...');
+      await Promise.all(files.map(item => fs.remove(item)));
+      console.log('清理dist目录完成');
+    }
     await this.clearOldVersionFiles();
   };
